Merged revisions 10881 via svnmerge from 
https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.8

........
  r10881 | dworley | 2007-10-05 17:06:34 -0400 (Fri, 05 Oct 2007) | 3 lines
  
  Patch from Sakeeb of HP:  Fixes so that sipX builds on HP-UX.
  Details in XECS-98.
........


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@10894 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXtackLib/include/resparse/types.h
@@ -33,7 +33,7 @@ typedef unsigned char boolean;
 #if ULONG_MAX == MAX32U
 typedef short int16;
 typedef int   int32;
-typedef unsigned long  u_int32;
+typedef unsigned int   u_int32;
 typedef unsigned short u_int16;
 /* 16 bit machines */
 #elif ULONG_MAX == 0xFFFF
@@ -48,7 +48,7 @@ typedef int   int32;
 typedef unsigned int   u_int32;
 typedef unsigned short u_int16;
 #endif
-typedef char int8;
+typedef signed char int8;
 typedef unsigned char u_int8;
 
 #endif /* RTP_VERSION */
